Move builder to main crate
