+++ red4ext/src/types.rs
@@ -5,7 +5,7 @@ pub use ffi::IScriptable;
 use red4ext_sys::ffi;
 pub use red4ext_sys::interop::{
     CName, EntityId, GameEItemIdFlag, GamedataItemStructure, ItemId, RaRef, RedString, ResRef,
-    ResourcePath, ResourcePathBuilder, TweakDbId, Variant, VoidPtr,
+    ResourcePath, TweakDbId, Variant, VoidPtr,
 };
 
 use crate::conv::{FromRepr, IntoRepr, NativeRepr};
